linux port

+++ srcs/get_next_line.c
@@ -23,7 +23,7 @@ static int
 		return (0);
 	while ((pnl = ft_strnlchr(tbuf->buf, tbuf->ofs)) == -1)
 	{
-		n = tbuf->ltr ? tbuf->ltr : BUFFER_SIZE - tbuf->ofs;
+		n = tbuf->ltr ? (size_t)tbuf->ltr : (BUFFER_SIZE - tbuf->ofs);
 		if (!(*line = ft_strxtd(*line, tbuf->buf + tbuf->ofs, n)))
 			return (-1);
 		tbuf->ofs = 0;
@@ -37,7 +37,7 @@ static int
 	if (n && n == (size_t)tbuf->ltr)
 		return (0);
 	if (tbuf->ltr)
-		tbuf->ltr = tbuf->ltr - (n + 1) == 0 ? -1 : tbuf->ltr - (n + 1);
+		tbuf->ltr = tbuf->ltr - (n + 1) == 0 ? (ssize_t)(-1) : (ssize_t)(tbuf->ltr - (n + 1));
 	tbuf->ofs += n + 1;
 	return (1);
 }
